ticket-monitor/requirements.txt
python-dotenv>=1.0.0
playwright>=1.42.0
APScheduler>=3.10.4
beautifulsoup4>=4.12.0

